MNT: rename control_layers/ to control_layer/
