+++ docs/source/control_layer/index.rst
@@ -1,0 +1,8 @@
+==============
+Control Layers
+==============
+.. automodule:: squirrel.control_layer.core
+   :members:
+
+.. automodule:: squirrel.control_layer.status
+   :members:
