--- docs/source/control_layers/index.rst
@@ -1,8 +1,0 @@
-==============
-Control Layers
-==============
-.. automodule:: squirrel.control_layers.core
-   :members:
-
-.. automodule:: squirrel.control_layers.status
-   :members:
