+++ squirrel/control_layer/_aioca.py
@@ -8,7 +8,7 @@ from aioca import CANothing, caget, camonitor, caput
 from aioca.types import AugmentedValue
 from epicscorelibs.ca import dbr
 
-from squirrel.control_layers._base_shim import _BaseShim
+from squirrel.control_layer._base_shim import _BaseShim
 from squirrel.errors import CommunicationError
 from squirrel.model import EpicsData, Severity, Status
