+++ squirrel/tests/test_ioc.py
@@ -1,4 +1,4 @@
-from squirrel.control_layers.core import ControlLayer
+from squirrel.control_layer.core import ControlLayer
 
 
 def test_ioc(linac_ioc):
